// src/api/models/agendamentoTipoEnum.ts
/**
 * Generated by orval v7.11.1 🍺
 * Do not edit manually.
 * MindMeet API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `online` - Online
* `presencial` - Presencial
 */
export type AgendamentoTipoEnum = typeof AgendamentoTipoEnum[keyof typeof AgendamentoTipoEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AgendamentoTipoEnum = {
  online: 'online',
  presencial: 'presencial',
} as const;