// src/api/models/estadoEnum.ts
/**
 * Generated by orval v7.11.1 🍺
 * Do not edit manually.
 * MindMeet API
 * API REST do **MindMeet**, uma plataforma de gestão para clínicas e consultórios de psicologia.

Permite o gerenciamento completo de psicólogos, pacientes, agendamentos e anotações clínicas, com suporte a dois modelos de operação:

- **Independente** — psicólogo autônomo que gerencia sua própria agenda e pacientes.
- **Institucional** — psicólogos e estagiários vinculados a uma instituição, supervisionados por um gestor.

A autenticação é feita via **JWT**. Todos os endpoints protegidos exigem o token no header `Authorization: Bearer <token>`.
 * OpenAPI spec version: 0.1.0
 */

/**
 * * `agendado` - Agendado
 * `cancelado` - Cancelado
 * `realizado` - Realizado
 */
export type EstadoEnum = (typeof EstadoEnum)[keyof typeof EstadoEnum]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EstadoEnum = {
  agendado: 'agendado',
  cancelado: 'cancelado',
  realizado: 'realizado',
} as const